Move assets into appropriate folders
