+++ frontend/src/css/algolia.css
