+++ frontend/src/js/audios.js
@@ -15,7 +15,7 @@ import {
   stats,
 } from 'instantsearch.js/es/widgets';
 import 'mdn-polyfills/Element.prototype.toggleAttribute';
-import './algolia.css';
+import '../css/algolia.css';
 import { itemTemplate } from './audio-item-template';
 import './player';
 import { searchBar } from './search-bar';
