+++ frontend/src/js/duration.mjs
