+++ frontend/src/js/search-bar.js
