+++ frontend/src/js/search-result-item.js
