+++ frontend/src/js/sound-quality-rating.mjs
