+++ frontend/src/public/favicon-16x16.png
